moved key-value filters to the same module, and left shims with deprecation notice
