+++ exporters/filters/key_value_filters.py
@@ -1,3 +1,5 @@
+import re
+
 from exporters.filters.base_filter import BaseFilter
 from exporters.utils import nested_dict_value
 
@@ -33,3 +35,29 @@ class KeyValueBaseFilter(BaseFilter):
         Should be overriden by derived classes implementing custom match.
         """
         raise NotImplementedError
+
+
+class KeyValueFilter(KeyValueBaseFilter):
+    """
+    Filter items depending on keys and values
+
+        - keys (list)
+            It is a list of dicts with the following structure: {"key": "value"}.
+            The filter will delete those items that do not contain a
+            key "key" or, if they do, that key is not the same as "value".
+    """
+    def _match_value(self, found, expected):
+        return found == expected
+
+
+class KeyValueRegexFilter(KeyValueBaseFilter):
+    """
+    Filter items depending on keys and values using regular expressions
+
+        - keys (list)
+            It is a list of dicts with the following structure: {"key": "regex"}.
+            The filter will delete those items that do not contain a
+            key "key" or, if they do, that key value does not match "regex".
+    """
+    def _match_value(self, found, expected):
+        return bool(re.match(expected, found))
